django-backend/requirements.txt
asgiref==3.6.0
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
Django==4.1.5
django-cors-headers==4.3.1
django-sslserver==0.22
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pycparser==2.21
PyJWT==2.7.0
python-dotenv==0.21.1
pytz==2022.7
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
